refactor: restructure pkg/utils for 300-line compliance
- Split paths.go (310→84 lines) into modular packages
- Created platform package for VS Code-specific operations (140 lines)
- Created files package for general utilities (103+62 lines)
- Split paths_test.go (301→66 lines) into focused test modules
- Maintain backward compatibility with re-exported functions
- All functionality preserved with improved organization
